Student-Performance-Prediction/backend-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
motor==3.6.0
pymongo==4.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.12
